Refactored back to TransactionForm
